index.js refactor
